(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -6,7 +6,15 @@ const controller = {}     // Objeto vazio
 
 controller.create = async function (req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end();
 
+    // Verifica se existe o campo "password" em "req.body".
+    // Caso positivo, geramos o hash da senha antes de enviá-la
+    // ao BD
+    // (12 na chamada a bcrypt.hash() corresponde ao número de
+    // passos de criptografia utilizados no processo)
     if (req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12);
     }
@@ -26,6 +34,10 @@ controller.create = async function (req, res) {
 
 controller.retrieveAll = async function (req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end();
+
     const result = await prisma.user.findMany(
       // Omite o campo "password" do resultado
       // por questões de segurança
@@ -45,6 +57,13 @@ controller.retrieveAll = async function (req, res) {
 
 controller.retrieveOne = async function (req, res) {
   try {
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!(req?.authUser?.is_admin ||
+      Number(req?.authUser?.id) === Number(req.params.id)))
+      return res.status(403).end();
+
     const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questões de segurança
@@ -67,7 +86,15 @@ controller.retrieveOne = async function (req, res) {
 
 controller.update = async function (req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end();
 
+    // Verifica se existe o campo "password" em "req.body".
+    // Caso positivo, geramos o hash da senha antes de enviá-la
+    // ao BD
+    // (12 na chamada a bcrypt.hash() corresponde ao número de
+    // passos de criptografia utilizados no processo)
     if (req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12);
     }
@@ -92,6 +119,10 @@ controller.update = async function (req, res) {
 
 controller.delete = async function (req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end();
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
